fix compiler warnings

+++ mysql_mitm.c
@@ -179,7 +179,7 @@ handle_auth_packet_from_client (struct conn_data *conn_data,
                                 const uv_buf_t * uv_buf, size_t nread)
 {
     char user[64];
-    int user_len;
+    size_t user_len;
     struct conn_data *conn_data_remote;
     struct destination *destination = NULL;
     char *mysql_server = NULL, *c, *i, *userptr;
